Adds toy chests and contracts
Fixes #95
