+++ frog-team/app/Http/Controllers/NewsletterController.php
@@ -4,12 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Services\Newsletter;
 use Illuminate\Http\Request;
+use Exception;
 use Illuminate\Validation\ValidationException;
 
 class NewsletterController extends Controller
 {
     public function __invoke(Newsletter $newsletter)
-    {
+    {   
         request()->validate(['email' => 'required|email']);
 
         try {
