+++ frog-team/app/Services/Newsletter.php
@@ -2,28 +2,7 @@
 
 namespace App\Services;
 
-use MailchimpMarketing\ApiClient;
-
-class Newsletter
+interface Newsletter
 {
-    
-    public function subscribe(string $email, string $list = null)
-    {
-        $list ??= config('services.mailchimp.lists.subscribers');
-
-        return $this->client()->lists->addListMember($list, [
-            'email_address' => $email,
-            'status' => 'subscribed'
-        ]);
-    }
-
-    protected function client() 
-    {
-        $mailchimp = new ApiClient();
-
-        return $mailchimp->setConfig([
-            'apiKey' => config('services.mailchimp.key'),
-            'server' => 'us21'
-        ]);
-    }
+    public function subscribe(string $email, string $list = null);
 }
